rename conda.common.os to conda.common._os to avoid conflict with built in
