+++ conda/cli/main_info.py
@@ -176,7 +176,7 @@ def get_info_dict(system=False):
         netrc_file=netrc_file,
     )
     if on_win:
-        from ..common.os.windows import is_admin_on_windows
+        from ..common._os.windows import is_admin_on_windows
         info_dict['is_windows_admin'] = is_admin_on_windows()
     else:
         info_dict['UID'] = os.geteuid()
