+++ conda/common/_os/windows.py
