+++ tests/common/os/test_windows.py
@@ -2,7 +2,7 @@
 from __future__ import absolute_import, division, print_function, unicode_literals
 
 from conda.common.compat import on_win
-from conda.common.os.windows import is_admin_on_windows
+from conda.common._os.windows import is_admin_on_windows
 
 
 def test_is_admin_on_windows():
